แก้ไขหน้า index.php
• ให้แสดงเฉพาะกระทู้ของผู้ใช้งานที่ไม่ได้ถูกแบน และถ้าผู้ใช้งานที่ถูกแบนเข้ามาที่หน้านี้จะไม่สามารถ
สร้างกระทู้ใหม่ได้ (ไม่มีปุ่มสร้างกระทู้ใหม่)

+++ index.php
@@ -64,7 +64,7 @@
                 </ul>
             </span>
             <?php
-                if(isset($_SESSION['id'])){
+                if(isset($_SESSION['id']) && $_SESSION['role'] != "b"){
                     echo "<a href='newpost.php' class='btn btn-success btn-sm' style='float: right;'><i class='bi bi-plus'></i> สร้างกระทู้ใหม่</button></a>";
                 }
             ?>
@@ -74,27 +74,30 @@
                 $conn=new PDO("mysql:host=localhost;dbname=webboard;charset=utf8","root","");
                 if (isset($_GET['category'])) {
                     $category = $_GET['category'];
-                    $sql="SELECT t3.name,t1.title,t1.id,t2.login,t1.post_date, t1.user_id FROM post as t1
+                    $sql="SELECT t3.name,t1.title,t1.id,t2.login,t1.post_date, t1.user_id, t2.role FROM post as t1
                         INNER JOIN user as t2 ON (t1.user_id=t2.id)
                         INNER JOIN category as t3 ON (t1.cat_id=t3.id) WHERE t1.cat_id = $category ORDER BY t1.post_date DESC";
                 }
                 else{
-                    $sql="SELECT t3.name,t1.title,t1.id,t2.login,t1.post_date, t1.user_id FROM post as t1
+                    $sql="SELECT t3.name,t1.title,t1.id,t2.login,t1.post_date, t1.user_id, t2.role FROM post as t1
                         INNER JOIN user as t2 ON (t1.user_id=t2.id)
                         INNER JOIN category as t3 ON (t1.cat_id=t3.id) ORDER BY t1.post_date DESC";
                 }
                 $result=$conn->query($sql);
                 while($row = $result->fetch()){
-                    echo "<tr><td class='d-flex justify-content-between'>
-                    <div>[ $row[0] ] <a href=post.php?id=$row[2]
-                    style=text-decoration:none>$row[1]</a><br>$row[3] - $row[4]</div>";
+                    if ($row[6] != 'b'){
+                        echo "<tr><td class='d-flex justify-content-between'>
+                        <div>[ $row[0] ] <a href=post.php?id=$row[2]
+                        style=text-decoration:none>$row[1]</a><br>$row[3] - $row[4]</div>";
+                    }
+                
                     if (isset($_SESSION['id'])){
-                        if ($_SESSION['user_id'] == $row[5]) {
+                        if ($_SESSION['user_id'] == $row[5] && $row[6] != 'b') {
                             echo "<div>
                                     <a href='delete.php?id={$row[2]}' class='btn btn-danger btn-sm float-end' onclick='return myFunction();'><i class='bi bi-trash'></i></a>
                                     <a href='editpost.php?id={$row[2]}' class='btn btn-warning btn-sm float-end me-2'><i class='bi bi-pencil'></i></a>
                                   </div>";
-                        } elseif ($_SESSION['role'] == "a") {
+                        } elseif ($_SESSION['role'] == "a" && $row[6] != 'b') {
                             echo "<div>
                                     <a href='delete.php?id={$row[2]}' class='btn btn-danger btn-sm float-end' onclick='return myFunction();'><i class='bi bi-trash'></i></a>
                                   </div>";
